Better separate session and direct-tcpip channels
