+++ go.mod
@@ -9,6 +9,6 @@ require (
 	github.com/sirupsen/logrus v1.8.1
 	github.com/valyala/fasthttp v1.26.0
 	golang.org/x/crypto v0.0.0-20210513164829-c07d793c2f9a
-	golang.org/x/term v0.0.0-20210503060354-a79de5458b56
+	golang.org/x/term v0.0.0-20210503060354-a79de5458b56 // indirect
 	gopkg.in/yaml.v2 v2.4.0
 )
